% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-web.R
\name{config_web}
\alias{config_web}
\title{write web configuration}
\usage{
config_web(file, folder, maxzoom, epsg, dates, formatdates, latIni, lonIni,
  latEnd, lonEnd, timeIni, timeEnd, varmin = 0, varmax = 100,
  legend = "NaN")
}
\arguments{
\item{folder}{folder}

\item{latIni}{latIni}

\item{lonIni}{lonIni}

\item{latEnd}{latEnd}

\item{lonEnd}{lonEnd}

\item{timeIni}{timeIni}

\item{timeEnd}{timeEnd}

\item{maxZoom}{maxZoom}
}
\description{
write web configuration
}
