% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{create_nc_name}
\alias{create_nc_name}
\title{Create new nc names for rechunked files and chunk directories}
\usage{
create_nc_name(file_name, sufix = "-t", ext = "")
}
\arguments{
\item{file_name}{Original netCDF file}

\item{sufix}{Sufix to put after bare filename and before extension}

\item{ext}{New file extension}
}
\description{
Create new nc names for rechunked files and chunk directories
}
\examples{
create_nc_name(file_name="ETo.nc", sufix="-t", ext=".bin")
}
