% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{get_struct_typecode}
\alias{get_struct_typecode}
\title{Converts the data type of a variable or dimension of a netCDF to the data type codes used
by the struct library.}
\usage{
get_struct_typecode(nc_type)
}
\arguments{
\item{nc_type}{netCDF data type}
}
\value{
struct data type
}
\description{
Converts the data type of a variable or dimension of a netCDF to the data type codes used
by the struct library.
}
\examples{
get_struct_typecode(nc_type = "double")
}
