% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getVarName}
\alias{getVarName}
\title{Get var name of main variable in nc file}
\usage{
getVarName(nc)
}
\arguments{
\item{nc}{open nc file}
}
\value{
var name
}
\description{
Get var name of main variable in nc file
}
