% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{maxFusion}
\alias{maxFusion}
\title{merge max arrays}
\usage{
maxFusion(max1, max2, positions)
}
\arguments{
\item{max1}{max1}

\item{max2}{max2}

\item{positions}{positions}
}
\value{
max
}
\description{
merge max arrays
}
