% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{merge_csv}
\alias{merge_csv}
\title{merge csv folders
transform file names so do not overlap}
\usage{
merge_csv(folders, folder)
}
\arguments{
\item{folders}{folders}

\item{folder}{folder}
}
\description{
merge csv folders
transform file names so do not overlap
}
\examples{
merge_csv(folders=c("pen", "bal", "can"), folder="merge")
}
