% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-layer.R
\name{merge_csv_layer}
\alias{merge_csv_layer}
\title{merge csv layers
transform index values so do not overlap}
\usage{
merge_csv_layer(folders, folder)
}
\arguments{
\item{folders}{folders}

\item{folder}{folder}
}
\description{
merge csv layers
transform index values so do not overlap
}
\examples{
merge_csv_layer(folders=c("pen", "bal", "can"), folder="merge")
}
