% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-layer.R
\name{merge_data_layer}
\alias{merge_data_layer}
\title{merge data layers}
\usage{
merge_data_layer(folders, folder)
}
\arguments{
\item{folders}{folders}

\item{folder}{folder}
}
\description{
merge data layers
}
\examples{
merge_data_layer(folders=c("pen", "bal", "can"), folder="merge")
}
