% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{minFusion}
\alias{minFusion}
\title{merge min arrays}
\usage{
minFusion(min1, min2, positions)
}
\arguments{
\item{min1}{min1}

\item{min2}{min2}

\item{positions}{positions}
}
\value{
min
}
\description{
merge min arrays
}
