% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncwebmapper-package.R
\docType{package}
\name{ncwebmapper}
\alias{ncwebmapper}
\alias{ncwebmapper-package}
\title{ncwebmapper}
\description{
ncwebmapper
}
\author{
Fergus Reig Gracia 
Maintainer: Fergus Reig Gracia
\url{https://github.com/ncwebmapper/ncwebmapper}
}
\keyword{package}
