% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tile_bounds}
\alias{tile_bounds}
\title{bounds of a tile}
\usage{
tile_bounds(nx, ny, zoom)
}
\arguments{
\item{nx}{nx}

\item{ny}{ny}

\item{zoom}{zoom}
}
\value{
tile bounds
}
\description{
bounds of a tile
}
