% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{write_csv}
\alias{write_csv}
\title{write csv files in a folder}
\usage{
write_csv(file, folder, epsg, dates, formatdates)
}
\arguments{
\item{file}{file}

\item{folder}{folder}

\item{epsg}{epsg}

\item{dates}{dates}

\item{formatdates}{formatdates}
}
\description{
write csv files in a folder
}
\examples{
write_csv(file="SPREAD_pen_pcp.nc", folder="pen")
write_csv(file="SPREAD_bal_pcp.nc", folder="bal")
write_csv(file="SPREAD_can_pcp.nc", folder="can")
}
