% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-layer.R
\name{write_csv_layer}
\alias{write_csv_layer}
\title{write csv query layer}
\usage{
write_csv_layer(file, folder, epsg, zoom)
}
\arguments{
\item{file}{file}

\item{folder}{folder}

\item{epsg}{epsg}

\item{zoom}{zoom}
}
\description{
write csv query layer
}
\examples{
write_csv_layer(file="SPREAD_pen_pcp.nc", folder="pen")
write_csv_layer(file="SPREAD_bal_pcp.nc", folder="bal")
write_csv_layer(file="SPREAD_can_pcp.nc", folder="can")
}
