% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-layer.R
\name{write_data_layer}
\alias{write_data_layer}
\title{write data layer}
\usage{
write_data_layer(file, folder, epsg, maxzoom, timeshift = 0)
}
\arguments{
\item{file}{file}

\item{folder}{folder}

\item{epsg}{epsg}

\item{maxzoom}{maxzoom}

\item{timeshift}{timeshift}
}
\description{
write data layer
}
\examples{
write_data_layer(file="SPREAD_pen_pcp.nc", folder="pen")
write_data_layer(file="SPREAD_bal_pcp.nc", folder="bal", timeshift=7670)
write_data_layer(file="SPREAD_can_pcp.nc", folder="can", timeshift=7670)
}
