% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{write_nc_chunk_t}
\alias{write_nc_chunk_t}
\title{Create the new netCDF file with chunk dimensions that favor obtaining temporal series
for each pixel.}
\usage{
write_nc_chunk_t(
  in_file,
  out_file,
  lon_by = -1,
  lat_by = -1,
  lon_name = "lon",
  lat_name = "lat",
  signif_digits
)
}
\arguments{
\item{in_file}{Original netCDF file}

\item{out_file}{netCDF file with the same information as the original but with new chunk structure.}

\item{lon_by}{Number of pixels horizontally that will be read as a block during the read/write loop. -1 to read all at once.}

\item{lat_by}{Number of pixels vertically that will be read as a block during the read/write loop. -1 to read all at once.}

\item{lon_name}{Name of longitude dimension.}

\item{lat_name}{Name of latitude dimension.}

\item{signif_digits}{Number of significant digits to keep in the data.}
}
\description{
Create the new netCDF file with chunk dimensions that favor obtaining temporal series
for each pixel.
}
\examples{
write_nc_chunk_t(in_file = "/path/ETo.nc", out_file = "/path/ETo-t.nc", lon_by = 100, lat_by = 100, lon_name = "lon", lat_name = "lat")
}
