% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{write_nc_t_chunk_dir}
\alias{write_nc_t_chunk_dir}
\title{Create the binary chunks directory for the nc oriented to the time series of each pixel.}
\usage{
write_nc_t_chunk_dir(in_file, out_file)
}
\arguments{
\item{in_file}{netCDF file with chunking oriented to the time series of each pixel.}

\item{out_file}{Chunks directory file.}
}
\description{
Create the binary chunks directory for the nc oriented to the time series of each pixel.
}
\examples{
write_nc_t_chunk_dir(in_file = "/path/ETo-t.nc", out_file = "/path/ETo-t.bin")
}
