% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{write_nc_xy_chunk_dir_iter}
\alias{write_nc_xy_chunk_dir_iter}
\title{Create the binary chunks directory for the nc oriented to the maps of each date
using the new H5Dchunk_iter function (https://docs.hdfgroup.org/hdf5/v1_14/group___h5_d.html#title6).}
\usage{
write_nc_xy_chunk_dir_iter(in_file, out_file)
}
\arguments{
\item{in_file}{netCDF file with chunking oriented to the maps of each date.}

\item{out_file}{Chunks directory file.}
}
\description{
Create the binary chunks directory for the nc oriented to the maps of each date
using the new H5Dchunk_iter function (https://docs.hdfgroup.org/hdf5/v1_14/group___h5_d.html#title6).
}
\examples{
write_nc_xy_chunk_dir_iter(in_file = "/path/ETo-xy.nc", out_file = "/path/ETo-xy.bin")
}
