% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{create_nc_name}
\alias{create_nc_name}
\title{Filename generator for rechunked nc and chunk directories.}
\usage{
create_nc_name(file_name, sufix = "-t", ext = "")
}
\arguments{
\item{file_name}{Original netCDF filename}

\item{sufix}{Optional suffix to be added to the end of the file name before the extension}

\item{ext}{File extension. If given the empty value, the original extension is maintained}
}
\value{
New filename
}
\description{
Filename generator for rechunked nc and chunk directories.
}
\examples{
create_nc_name(file_name = "ETo.nc", sufix = "-t", ext = ".bin")
}
