% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{fusion_pen_can}
\alias{fusion_pen_can}
\title{Union of peninsular and Canary Islands netCDF files.}
\usage{
fusion_pen_can(
  can_filename,
  pen_filename,
  fusion_filename,
  nc_chunk_num = 16,
  signif_digits,
  lon_name = "lon",
  lat_name = "lat",
  time_name = "time",
  grid_size
)
}
\arguments{
\item{can_filename}{netCDF file with data for the Canary Islands.}

\item{pen_filename}{netCDF file with data for the Iberian Peninsula.}

\item{fusion_filename}{netCDF file with the union of the data from the two files.}

\item{nc_chunk_num}{Number of chunks to use in the new netCDF file in each dimension.}

\item{signif_digits}{Number of significant digits to keep in the data.}

\item{lon_name}{Name of longitude dimension.}

\item{lat_name}{Name of latitude dimension.}

\item{time_name}{Name of time dimension.}

\item{grid_size}{Size of the grid in degrees.}
}
\description{
Union of peninsular and Canary Islands netCDF files.
}
\examples{
fusion_pen_can(can_filename = "/path/ETo_can.nc", pen_filename = "/path/ETo_pen.nc", fusion_filename = "/path/ETo_all.nc", nc_chunk_num = 16, signif_digits = 4)
}
